+++ GooglePay/src/main/java/com/braintreepayments/api/PaymentMethodNonceFactory.java
@@ -5,15 +5,15 @@ import org.json.JSONObject;
 
 import java.util.Iterator;
 
-public class PaymentMethodNonceFactory {
-    public static JSONObject extractPaymentMethodToken(String paymentDataString) throws JSONException {
+class PaymentMethodNonceFactory {
+    static JSONObject extractPaymentMethodToken(String paymentDataString) throws JSONException {
         return new JSONObject(new JSONObject(paymentDataString)
                         .getJSONObject("paymentMethodData")
                         .getJSONObject("tokenizationData")
                         .getString("token"));
     }
 
-    public static PaymentMethodNonce fromString(String paymentDataString) throws JSONException {
+    static PaymentMethodNonce fromString(String paymentDataString) throws JSONException {
         JSONObject token = extractPaymentMethodToken(paymentDataString);
         Iterator<String> keys = token.keys();