+++ SharedUtils/src/main/java/com/braintreepayments/api/ServiceUnavailableException.java
@@ -4,9 +4,9 @@ package com.braintreepayments.api;
  * Exception thrown when a 503 HTTP_UNAVAILABLE response is encountered. Indicates the server is
  * unreachable or the request timed out.
  */
-public class DownForMaintenanceException extends Exception {
+public class ServiceUnavailableException extends Exception {
 
-    DownForMaintenanceException(String message) {
+    ServiceUnavailableException(String message) {
         super(message);
     }
 }