+++ VisaCheckout/src/main/java/com/braintreepayments/api/VisaCheckoutAccount.java
@@ -8,7 +8,7 @@ import org.json.JSONObject;
 /**
  * Use to construct a Visa Checkout tokenization request.
  */
-public class VisaCheckoutAccount extends PaymentMethod {
+class VisaCheckoutAccount extends PaymentMethod {
     private static final String CALL_ID = "callId";
     private static final String ENCRYPTED_KEY = "encryptedKey";
     private static final String ENCRYPTED_PAYMENT_DATA = "encryptedPaymentData";
@@ -21,7 +21,7 @@ public class VisaCheckoutAccount extends PaymentMethod {
     /**
      * @param visaPaymentSummary returned from Visa Checkout after a successful payment.
      */
-    public VisaCheckoutAccount(VisaPaymentSummary visaPaymentSummary) {
+    VisaCheckoutAccount(VisaPaymentSummary visaPaymentSummary) {
         if (visaPaymentSummary == null) {
             return;
         }